module renaming to conform to python style guide
