+++ meta-my-iotc-python-sdk-example/recipes-apps/iotc-telemetry-demo/files/model/enums.py
