+++ meta-my-iotc-python-sdk-example/recipes-apps/iotc-telemetry-demo/files/model/json_parser.py
